apps/backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1

# Storage and media handling
boto3==1.35.0
Pillow==11.0.0
qrcode==7.4.2
python-magic==0.4.27

pytest
pytest-django
